backend/requirements.txt
django
psycopg2-binary
requests
djangorestframework
dj-rest-auth
django-allauth
django-cors-headers
djangorestframework-simplejwt
python-decouple
ollama
strands-agents
strands-agents-tools
channels
channels-redis
asyncio
daphne
minio